Fix app namespace issue
Problem:
Apps of the same name deploy into the same namespace. This fails when a
namespace is owned by another project

Solution:
Apps should be deployed into a unique namespace unless specified
otherwise.
Add a flag to specify the namespace
Create a namespace with a random suffix to deploy the app into if
namespace is not specified