Add message when project/cluster unavailable.
Previously, both functions would return 404 with no additional information or recommended actions.
Now, both inform the user that either the current project or cluster is not available, respectively.
It is then suggested the user runs `rancher context switch`. Finally, the error is displayed.
This will aid the user in resolving the error.