Call yaml.Unmarshal on values/answers file using YAML
We use json.Unmarshal for files using JSON format, and convert the
YAML format files to JSON first and call json.Unmarshal on them as well.
The yaml to json conversion let to unquoted numbers getting converted to
scientific notation. So this commit fixes it by calling yaml.Unmarshal
directly on file with yaml format instead of converting them to json.