Support SSH for v2 provisioned machines
Before this change, it was not possible to use the SSH command to
connect to machines provisioned with v2 provisioning. After this change
(and including the changes to Rancher), the CLI will use the new CAPI
client to get the SSH key and config from Rancher for v2 provisioned
machines.

A side effect of this change is the addition of the new `rancher
machines ls` command that lists all machines for the current cluster
context.